requirements.txt
apache-airflow==2.9.0
pandas==2.2.2
requests==2.31.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pytest==8.2.0